Allow empty key properties (#16)
* Fix #15 - allow empty key_properties

The issue was that we were always adding a 'required' property onto the schema for the records, to make the properties listed as key_properties required. When there were no key properties, we were setting 'required' to an empty array, which apparently is not allowed in JSON-schema. I just changed it so we don't set 'required' if there are no key_properties. Added a test to cover it and functional-tested by piping the 'empty_key_properties.json' file into target-stitch and confirming that it persisted the record to Stitch.

* Add sample data for empty key properties